dev 1.9
Debugged the router, added a main page to it and implemented the projects router. Added the react-parallax effect on the about page and created the grid for the pictured. initially styled the skills section and added initial content on the grad, ps2 and grad design pages. moreover, redesigned the footer buttons